Load Scene Resources Relative to Scene File
Resources or other files referenced in scenes will be loaded relative to
scene file. This may not implement proper loading of all resources yet
but it works for meshes and materials.

This allows scenes to be loaded from subdirectories of the program root
directory with the purpose of allowing modding support.