Add extra functionality to the Std Math
I noticed the Math class was deleted because lot of functions already existed in std Math, but maybe we can keep some extra functions that are not in the std Math? Having these extras may help a lot not to have to code so common functions manually every time they are needed.