Add support for shadow map atlasing
- setTargetStream(), drawMeshesStream(), endStream()

These functions were created as an alternative to allow rendering continuously to the same render target, hence the "Stream" in the names.
They are modified versions of setTarget() and drawMeshes(), with some differences but the most important is that the logic that ended
the rendering process was moved to a function, so they require to manually call endStream() to close it properly.
What motivated the creation of these functions is that it was observed that drawMeshes() closed the rendering process for you when
you called it, which is ok when rendering to a normal render target, but for a render target that is supposedly being rendered over and over,
this posed some overhead that could be solved by manually allowing closing the process.

drawMeshesStream() also removed some of the logic from the original, specifically the logic to detect if a light should be skipped,
since this is already done in Inc so it was wasted computing, and the logic to configure the viewport for cascading,
because it didn't make sense to couple this class with atlasing logic.

- zToShadowMapScale(), shadowMapScale, culledLight

These are some of the few changes introduced to `updateClusters()` in order to support LOD and view culling of lights shadows for atlasing.
View culling works by detecting incorrect states of clustering which are deemed to result when lights fall out of view.

LOD supports works by taking advantage of the clustering algorithm, which uses the minZ value when computing clusters in zToShadowMapScale() to
determine the size of the shadow map for a light arbitrarily.

zToShadowMapScale() is a custom function that plots from [1, 16] (cluster Z slices) to [1.0, 0.0] (shadow map scale). Good enough and cheaper results
were observed than if they were implemented with the solution proposed in the clustering shadows algorithm (determine the angle for each cluster).

- discardLight(), discardLightCulled()

These functions were added to make more obvious the dependency of looping (and discarding) lights under the same exact conditions,
otherwise is prone to issues because the uniform data for cluster is different from the lights array data. Probably a better solution would be to maybe centralize
looping instead of just the conditions to discard lights, but it would require some thought since there are a few different loops for clustering than for lights array

- pointLightsData

this array is updated by Inc from armory, and it's used to provide data to compute UV for point lights in the atlas. It didn't make sense to couple
Uniforms nor LightObject with atlasing logic, so it's just in LightObject to avoid coupling.

- updateLWVPMatrixArray(), LWVPMatrixArray

Allow passing arrays of uniform world view projection matrices. The uv coordinates of the specific tiles in the atlas
is composed directly into the transformation, so it saves gpu cycles.