Added the extra inversion to spot light matrix multiplication for atlas
By doing this extra multiplication we avoid wasting gpu cycles to do it in the shader.
This should make sun work too since the same it's applied for cascades.
Also modified conditions of the bias matrix based on what was talked here https://github.com/armory3d/armory/issues/2110,
so it will the two extra inversions so it fits this model https://github.com/armory3d/armory/issues/2110#issuecomment-785433567 and work
in Firefox and Chrome (in Linux and Windows at least).