Moved generating biasLWVP matrix for sun to its own uniform case
This was done to separate a generic biasLWVP matrix from the one for
the sun in deferred light, by making an specific case of uniform for it.

The new uniform case will find the first directional light and use it to
generate the matrix, instead of relying on the user for setting
"renderpath.light" beforehand. This avoids issues.

_biasLightWorldViewProjectionMatrix was reverted to its original state
before shadow map atlases, so it can be used normally by users, without
requiring fiddling with variables used for atlasing.